pyproject.toml
[project]
name = "my-voice-bot"
version = "0.1.0"
description = "Voice AI bot built with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0,<1",
    "pipecat-ai[openai,runner,silero,livekit]",
    "uvicorn>=0.30.0,<1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]